pyproject.toml
[build-system]
requires = ["setuptools >= 69.5.1", "setuptools_scm[toml]>=8.1.0"]
build-backend = "setuptools.build_meta"

[project]
name = "jade"
license = {file = "LICENSE"}
dynamic = ["version"]
authors = [
  { name="JADE dev team", email="sc-radiationtransport@f4e.europa.eu" },
]
description = "JADE, a V&V tool for nuclear data libraries and transport codes"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: European Union Public Licence 1.2 (EUPL 1.2)",
    "Operating System :: OS Independent",
]
dependencies = [
    "pandas",
    "numjuggler",
    "tqdm",
    "xlsxwriter",
    "openpyxl",
    "matplotlib",
    # Latest version of scipy that supports Python 3.9.
    # This is a temporary fix and should be removed once Python 3.9 support is dropped in JADE. 
    "scipy == 1.13.1",
    "python-docx",
    "aspose-words",
    "requests",
    "f4enix >= 0.8.1",
#    openmc @ git+https://github.com/openmc-dev/openmc.git@v0.14.0#egg=openmc
    "pyyaml",
    "ttkthemes",
]
[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "pytest-xvfb",  # needed for headless display
    "ruff"  # best formatter that should be used by everybody
    ]
openmc = [
    "openmc @ git+https://github.com/openmc-dev/openmc.git@v0.14.0#egg=openmc"
    ]

[project.urls]
Homepage = "https://github.com/JADE-V-V/JADE"
Repository = "https://github.com/JADE-V-V/JADE"
Documentation = "https://jade-a-nuclear-data-libraries-vv-tool.readthedocs.io/en/latest/"

# [project.scripts]
# jade = "jade.main:main"

[tool.setuptools_scm]
write_to = "src/_version.py"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["jade"]

[tool.setuptools.package-data]
"*" = ["*.txt", '*.csv', '*.sh', '*.cmd']

